Dysplasia detection rate on surveillance

This function assesses the dysplasia detection rate per endoscopist.

BarrettsDDRQual(dataframe, EndoscopistReportColumn, IMorNoIM)

Arguments

dataframe
dataframe,

EndoscopistReportColumn
column with the Endoscopist names

IMorNoIM
extracted with the function Barretts_PathStage()

Examples

# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
v<-HistolDx(Mypath,'Diagnosis')

#> Warning: provided 10 variables to replace 1 variables

v<-HistolExtrapolDx(v,'Diagnosis')
v<-HistolNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
#The function relies on the other Barrett's functions being run as well:
b1<-Barretts_PragueScore(v,'Findings')
b2<-Barretts_PathStage(b1,'Histology')
b3<-Barretts_EventType(b2,'Histology',
'ProcedurePerformed','Indications','Findings')
# The follow-up group depends on the histology and the Prague score for a
# patient so it takes the processed Barrett's data and then looks in the
# Findings column for permutations of the Prague score.
b4<-Barretts_FUType(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
# The function takes the column with the extracted worst grade of
# histopathology and returns the proportion of each finding (ie
# proportion with low grade dysplasia, high grade etc.) for each
# endoscopist
BarrettsDDRQual(b4,'Endoscopist','IMorNoIM')

#>                                  
#>                                            HGD           IM          LGD
#>    Dr\n al-Arif, Ummu Kulthoom\n  0.0071090047 0.0208530806 0.0047393365
#>    Dr\n Anderson, Alana\n         0.0061611374 0.0213270142 0.0037914692
#>    Dr\n Avitia-Ramirez, Alondra\n 0.0061611374 0.0208530806 0.0052132701
#>    Dr\n Greimann, Phoua\n         0.0047393365 0.0165876777 0.0037914692
#>    Dr\n Ives, Rashiah\n           0.0056872038 0.0232227488 0.0037914692
#>    Dr\n Kekich, Annabelle\n       0.0052132701 0.0180094787 0.0023696682
#>    Dr\n Kola-Kehinde, Karisma\n   0.0075829384 0.0161137441 0.0075829384
#>    Dr\n Martinez, Maegen\n        0.0090047393 0.0189573460 0.0042654028
#>    Dr\n Moreno, Lauren\n          0.0071090047 0.0180094787 0.0104265403
#>    Dr\n Sullivan, Shelby\n        0.0094786730 0.0213270142 0.0066350711
#>                                  
#>                                          No_IM          T1a
#>    Dr\n al-Arif, Ummu Kulthoom\n  0.0777251185 0.0018957346
#>    Dr\n Anderson, Alana\n         0.0725118483 0.0023696682
#>    Dr\n Avitia-Ramirez, Alondra\n 0.0616113744 0.0014218009
#>    Dr\n Greimann, Phoua\n         0.0635071090 0.0004739336
#>    Dr\n Ives, Rashiah\n           0.0687203791 0.0009478673
#>    Dr\n Kekich, Annabelle\n       0.0649289100 0.0000000000
#>    Dr\n Kola-Kehinde, Karisma\n   0.0654028436 0.0009478673
#>    Dr\n Martinez, Maegen\n        0.0663507109 0.0004739336
#>    Dr\n Moreno, Lauren\n          0.0744075829 0.0004739336
#>    Dr\n Sullivan, Shelby\n        0.0592417062 0.0004739336

rm(v)