Analysis of Barrett's Documentation Quality

This function assesses the Barrett's documentation. This notes how many reports contain the mandatory report fields as specified in the BSG standards on Barrett's endoscopic reporting. This should be run after the BarrettsDataAccord_Prague.

BarrettsQuality_AnalysisDocumentation(x, Findings)

Arguments

x
dataframe

Findings
column of interest- usually the main body of the endoscopic report

Examples

# Firstly relevant columns are extrapolated from the
b1<-BarrettsDataAccord_Prague(Myendo,'Findings')
# The documentation is really from the endoscopic Findings column
BarrettsQuality_AnalysisDocumentation(b1,'Findings')

#> Error in barchart(b ~ Proportion, data = EndoMinDataSet, groups = s, scales = list(x = list(rot = 0,     cex = 2), y = list(cex = 2), main = list("Endoscopy documentation for Barrett's PRE DOI",     cex = 2.8)), ): could not find function "barchart"

rm(v)

#> Warning: object 'v' not found