Unique Hospital Numbers of Barrett's patients

This function gets the unique patient ID's for each patient, for each rule. It lists the unique PatientIDs assocaited with a rule ('Rule1','Rule2','Rule3','NoRules'). This allows us to determine how many patients will need follow up at specific time intervals. This should be run after the Barretts_PragueScore and Barretts_PathStage.

BarrettsSurveil_HospNum(dataframe, rule, PatientID)

Arguments

dataframe
dataframe with column of interest

rule
Rule of interest

PatientID
Column containing patient numbers

Examples

# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
v<-HistolAccessionNumber(Mypath,'Histology',
'SP-\\d{2}-\\d{7}')
v<-HistolDx(v,'Diagnosis')
v<-HistolExtrapolDx(v,'Diagnosis')
v<-HistolNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
# The function relies on the other Barrett's functions being run as well:
b1<-Barretts_PragueScore(v,'Findings')
b2<-Barretts_PathStage(b1,'Histology')
b3<-Barretts_EventType(b2,'Histology',
'ProcedurePerformed','Indications','Findings')
# The follow-up group depends on the histology and the Prague score for a
# patient so it takes the processed Barrett's data and then looks in the
# Findings column for permutations of the Prague score.
b4<-Barretts_FUType(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
#Finally the unique hospital numbers are obtained according to the follow-up
# rule you are looking for
BarrettsSurveil_HospNum(b4,'Rule1','HospitalNumber')

#>               x
#> 1    J6044658\n
#> 2    Y6417773\n
#> 3    B6072011\n
#> 4    G1449886\n
#> 5    O7163832\n
#> 6    L4378217\n
#> 7    K2657390\n
#> 8    N4378127\n
#> 9    R8004923\n
#> 10   M5148114\n
#> 11   X2813264\n
#> 12   A6871788\n
#> 13   J6229707\n
#> 14   Y7228898\n
#> 15   B1224656\n
#> 16   O4756019\n
#> 17   P2080807\n
#> 18   L2478495\n
#> 19   S1951882\n
#> 20   N2752405\n
#> 21   Q4907331\n
#> 22   R9120543\n
#> 23   F7095078\n
#> 24   T1732455\n
#> 25   O3104214\n
#> 26   L8917541\n
#> 27   D2568207\n
#> 28   N1257699\n
#> 29   H7489214\n
#> 30   Z8946728\n
#> 31   Y8211489\n
#> 32   I3619361\n
#> 33   W7335845\n
#> 34   O1085643\n
#> 35   S1985702\n
#> 36   D4796018\n
#> 37   Q7729897\n
#> 38   H9801234\n
#> 39   F7134775\n
#> 40   I8031481\n
#> 41   D5975610\n
#> 42   Z2727641\n
#> 43   U1759838\n
#> 44   C5917931\n
#> 45   K2068433\n
#> 46   Z6299612\n
#> 47   E6197772\n
#> 48   M8823141\n
#> 49   J1337672\n
#> 50   G8210675\n
#> 51   I1303991\n
#> 52   W2075882\n
#> 53   F7058648\n
#> 54   E7331206\n
#> 55   A3962018\n
#> 56   J8367396\n
#> 57   G9989010\n
#> 58   V1083206\n
#> 59   M9440150\n
#> 60   V1607560\n
#> 61   W2120051\n
#> 62   S9393699\n
#> 63   H4452093\n
#> 64   C8544445\n
#> 65   X5917376\n
#> 66   B8880712\n
#> 67   P4425100\n
#> 68   K6126215\n
#> 69   Q7176341\n
#> 70   R3882435\n
#> 71   T1294090\n
#> 72   G4762980\n
#> 73   V1418327\n
#> 74   I6545050\n
#> 75   U5272217\n
#> 76   A1648588\n
#> 77   V4085520\n
#> 78   S6740918\n
#> 79   U5156770\n
#> 80   B9209688\n
#> 81   P6658276\n
#> 82   L4154167\n
#> 83   E9204095\n
#> 84   Q7597311\n
#> 85   E7461659\n
#> 86   W1047941\n
#> 87   D9082891\n
#> 88   T9453315\n
#> 89   Z9871480\n
#> 90   P6620949\n
#> 91   H7140324\n
#> 92   C2789586\n
#> 93   F7957976\n
#> 94   Y8594667\n
#> 95   M4134709\n
#> 96   X5990901\n
#> 97   K8099671\n
#> 98   N2905244\n
#> 99   R2764793\n
#> 100  T3145793\n

rm(v)