Extracts the columns from the raw report

This is the main extractor for the Endoscopy and Histology report. This depends on the user creating a list of words or characters that act as the words that should be split against. The list is then fed to the Extractor in a loop so that it acts as the beginning and the end of the regex used to split the text. Whatever has been specified in the list is used as a column header. Column headers don't tolerate special characters like : or ? and / and don't allow numbers as the start character so these have to be dealt with in the text before processing

Extractor2(dataframeIn, Column, delim)

Arguments

dataframeIn
the dataframe

Column
the column to extract from

delim
the vector of words that will be used as the boundaries to extract against

x
the dataframe

y
the column to extract from

stra
the start of the boundary to extract

strb
the end of the boundary to extract

t
the column name to create

Examples

# As column names cant start with a number, one of the dividing
# words has to be converted
Myendo$OGDReportWhole<-gsub('2nd Endoscopist:','Second endoscopist:',
Myendo$OGDReportWhole)
# A list of dividing words (which will also act as column names)
# is then constructed
EndoscTree<-list('Hospital Number:','Patient Name:','General Practitioner:',
'Date of procedure:','Endoscopist:','Second Endoscopist:','Medications',
'Instrument','Extent of Exam:','Indications:','Procedure Performed:',
'Findings:','Endoscopic Diagnosis:')
# The Extractor function is then used as part of a loop to divide the raw
# report up according to the user-defined words in the extraction list
# defined above
for(i in 1:(length(EndoscTree)-1)) {
 Myendo<-Extractor(Myendo,'OGDReportWhole',as.character(EndoscTree[i]),
 as.character(EndoscTree[i+1]),as.character(EndoscTree[i]))
}
res<-Myendo
Extracts the columns from the raw report

This is the main extractor for the Endoscopy and Histology report.
This depends on the user creating a list of words or characters that
act as the words that should be split against. The list is then fed to the
Extractor in a loop so that it acts as the beginning and the end of the
regex used to split the text. Whatever has been specified in the list
is used as a column header. Column headers don't tolerate special characters
like : or ? and / and don't allow numbers as the start character so these
have to be dealt with in the text before processing

# As column names cant start with a number, one of the dividing
# words has to be converted
Myendo$OGDReportWhole<-gsub('2nd Endoscopist:','Second endoscopist:',
Myendo$OGDReportWhole)
# A list of dividing words (which will also act as column names)
# is then constructed
Extractor(Mypath,PathReportWhole,mywords)
res<-Myendo

#> Error: <text>:19:10: unexpected symbol
#> 18: res<-Myendo
#> 19: Extracts the
#>              ^