Extract the number of biopsies taken from histology report

This extracts the number of biopsies taken from the pathology report. This is usually from the Macroscopic description column. It collects everything from the regex [0-9]1,2.0,3 to whatever the string boundary is (z).

HistolNumbOfBx(dataframe, MacroColumn, regString)

Arguments

dataframe
the dataframe

MacroColumn
Column containing the Macroscopic description text

regString
The keyword to remove and to stop at in the regex

Examples

qq<-HistolNumbOfBx(Mypath,'Macroscopicdescription','specimen')