Determine overall number of procedures performed

Determines the number of endoscopies done by an endoscopist by type of endosopy and indication for a given timeframe As per BSG recommendations for Upper GI minimum number of gastroscopies in a year (although here the time frame is user defined)

NumberPerformed(dataframe, EndoscopistColumn, IndicationColumn)

Arguments

dataframe
The dataframe

EndoscopistColumn
The column containing the Endoscopists names

IndicationColumn
The column containing the Indication for the examination

Examples

Myendo<-Myendo[grepl('Gastroscopy',Myendo$ProcedurePerformed),]
 NumberPerformed(Myendo,'Endoscopist','Indications')

#>                        Var1                                        Var2 Freq
#> 1    al Arif, Ummu Kulthoom         chronic abdo pain and constipaton\n   17
#> 2            Anderson,Alana         chronic abdo pain and constipaton\n   10
#> 3   Avitia Ramirez, Alondra         chronic abdo pain and constipaton\n   10
#> 4            Greimann,Phoua         chronic abdo pain and constipaton\n   12
#> 5              Ives,Rashiah         chronic abdo pain and constipaton\n   16
#> 6          Kekich,Annabelle         chronic abdo pain and constipaton\n   11
#> 7     Kola Kehinde, Karisma         chronic abdo pain and constipaton\n   10
#> 8           Martinez,Maegen         chronic abdo pain and constipaton\n   13
#> 9             Moreno,Lauren         chronic abdo pain and constipaton\n   15
#> 10          Sullivan,Shelby         chronic abdo pain and constipaton\n   13
#> 11   al Arif, Ummu Kulthoom                                 Dyspepsia\n   18
#> 12           Anderson,Alana                                 Dyspepsia\n   13
#> 13  Avitia Ramirez, Alondra                                 Dyspepsia\n    8
#> 14           Greimann,Phoua                                 Dyspepsia\n   11
#> 15             Ives,Rashiah                                 Dyspepsia\n   15
#> 16         Kekich,Annabelle                                 Dyspepsia\n   12
#> 17    Kola Kehinde, Karisma                                 Dyspepsia\n   16
#> 18          Martinez,Maegen                                 Dyspepsia\n   12
#> 19            Moreno,Lauren                                 Dyspepsia\n   18
#> 20          Sullivan,Shelby                                 Dyspepsia\n   17
#> 21   al Arif, Ummu Kulthoom   Dysphagia/intermittent for a few months\n   10
#> 22           Anderson,Alana   Dysphagia/intermittent for a few months\n   20
#> 23  Avitia Ramirez, Alondra   Dysphagia/intermittent for a few months\n    5
#> 24           Greimann,Phoua   Dysphagia/intermittent for a few months\n   14
#> 25             Ives,Rashiah   Dysphagia/intermittent for a few months\n   13
#> 26         Kekich,Annabelle   Dysphagia/intermittent for a few months\n    7
#> 27    Kola Kehinde, Karisma   Dysphagia/intermittent for a few months\n   13
#> 28          Martinez,Maegen   Dysphagia/intermittent for a few months\n   12
#> 29            Moreno,Lauren   Dysphagia/intermittent for a few months\n   11
#> 30          Sullivan,Shelby   Dysphagia/intermittent for a few months\n   13
#> 31   al Arif, Ummu Kulthoom                     Dysphagia/Odynophagia\n   13
#> 32           Anderson,Alana                     Dysphagia/Odynophagia\n   13
#> 33  Avitia Ramirez, Alondra                     Dysphagia/Odynophagia\n    8
#> 34           Greimann,Phoua                     Dysphagia/Odynophagia\n   11
#> 35             Ives,Rashiah                     Dysphagia/Odynophagia\n   15
#> 36         Kekich,Annabelle                     Dysphagia/Odynophagia\n   12
#> 37    Kola Kehinde, Karisma                     Dysphagia/Odynophagia\n    7
#> 38          Martinez,Maegen                     Dysphagia/Odynophagia\n   12
#> 39            Moreno,Lauren                     Dysphagia/Odynophagia\n   19
#> 40          Sullivan,Shelby                     Dysphagia/Odynophagia\n    9
#> 41   al Arif, Ummu Kulthoom                   Follow-up ULCER HEALING\n    4
#> 42           Anderson,Alana                   Follow-up ULCER HEALING\n   11
#> 43  Avitia Ramirez, Alondra                   Follow-up ULCER HEALING\n   16
#> 44           Greimann,Phoua                   Follow-up ULCER HEALING\n   12
#> 45             Ives,Rashiah                   Follow-up ULCER HEALING\n   12
#> 46         Kekich,Annabelle                   Follow-up ULCER HEALING\n   15
#> 47    Kola Kehinde, Karisma                   Follow-up ULCER HEALING\n   14
#> 48          Martinez,Maegen                   Follow-up ULCER HEALING\n    9
#> 49            Moreno,Lauren                   Follow-up ULCER HEALING\n   10
#> 50          Sullivan,Shelby                   Follow-up ULCER HEALING\n   15
#> 51   al Arif, Ummu Kulthoom          Haematemesis or Melaena/Blood PR\n   17
#> 52           Anderson,Alana          Haematemesis or Melaena/Blood PR\n   13
#> 53  Avitia Ramirez, Alondra          Haematemesis or Melaena/Blood PR\n    9
#> 54           Greimann,Phoua          Haematemesis or Melaena/Blood PR\n    7
#> 55             Ives,Rashiah          Haematemesis or Melaena/Blood PR\n    9
#> 56         Kekich,Annabelle          Haematemesis or Melaena/Blood PR\n    4
#> 57    Kola Kehinde, Karisma          Haematemesis or Melaena/Blood PR\n   13
#> 58          Martinez,Maegen          Haematemesis or Melaena/Blood PR\n    6
#> 59            Moreno,Lauren          Haematemesis or Melaena/Blood PR\n   12
#> 60          Sullivan,Shelby          Haematemesis or Melaena/Blood PR\n   11
#> 61   al Arif, Ummu Kulthoom                    Nausea and/or Vomiting\n    9
#> 62           Anderson,Alana                    Nausea and/or Vomiting\n    7
#> 63  Avitia Ramirez, Alondra                    Nausea and/or Vomiting\n   13
#> 64           Greimann,Phoua                    Nausea and/or Vomiting\n    7
#> 65             Ives,Rashiah                    Nausea and/or Vomiting\n   13
#> 66         Kekich,Annabelle                    Nausea and/or Vomiting\n    6
#> 67    Kola Kehinde, Karisma                    Nausea and/or Vomiting\n    7
#> 68          Martinez,Maegen                    Nausea and/or Vomiting\n   11
#> 69            Moreno,Lauren                    Nausea and/or Vomiting\n   10
#> 70          Sullivan,Shelby                    Nausea and/or Vomiting\n    7
#> 71   al Arif, Ummu Kulthoom                     Oesophagus- Dysplasia\n    7
#> 72           Anderson,Alana                     Oesophagus- Dysplasia\n   18
#> 73  Avitia Ramirez, Alondra                     Oesophagus- Dysplasia\n    8
#> 74           Greimann,Phoua                     Oesophagus- Dysplasia\n    8
#> 75             Ives,Rashiah                     Oesophagus- Dysplasia\n   17
#> 76         Kekich,Annabelle                     Oesophagus- Dysplasia\n   13
#> 77    Kola Kehinde, Karisma                     Oesophagus- Dysplasia\n   19
#> 78          Martinez,Maegen                     Oesophagus- Dysplasia\n   12
#> 79            Moreno,Lauren                     Oesophagus- Dysplasia\n   11
#> 80          Sullivan,Shelby                     Oesophagus- Dysplasia\n   16
#> 81   al Arif, Ummu Kulthoom                                    Other-\n   32
#> 82           Anderson,Alana                                    Other-\n   24
#> 83  Avitia Ramirez, Alondra                                    Other-\n   23
#> 84           Greimann,Phoua                                    Other-\n   21
#> 85             Ives,Rashiah                                    Other-\n   21
#> 86         Kekich,Annabelle                                    Other-\n   19
#> 87    Kola Kehinde, Karisma                                    Other-\n   20
#> 88          Martinez,Maegen                                    Other-\n   25
#> 89            Moreno,Lauren                                    Other-\n   18
#> 90          Sullivan,Shelby                                    Other-\n   18
#> 91   al Arif, Ummu Kulthoom               Previous OGD ? 8 months ago\n   18
#> 92           Anderson,Alana               Previous OGD ? 8 months ago\n    9
#> 93  Avitia Ramirez, Alondra               Previous OGD ? 8 months ago\n   18
#> 94           Greimann,Phoua               Previous OGD ? 8 months ago\n   11
#> 95             Ives,Rashiah               Previous OGD ? 8 months ago\n   12
#> 96         Kekich,Annabelle               Previous OGD ? 8 months ago\n    7
#> 97    Kola Kehinde, Karisma               Previous OGD ? 8 months ago\n    9
#> 98          Martinez,Maegen               Previous OGD ? 8 months ago\n   12
#> 99            Moreno,Lauren               Previous OGD ? 8 months ago\n   15
#> 100         Sullivan,Shelby               Previous OGD ? 8 months ago\n   14
#> 101  al Arif, Ummu Kulthoom  Reflux-like Symptoms/Atypical Chest Pain\n    9
#> 102          Anderson,Alana  Reflux-like Symptoms/Atypical Chest Pain\n   14
#> 103 Avitia Ramirez, Alondra  Reflux-like Symptoms/Atypical Chest Pain\n   11
#> 104          Greimann,Phoua  Reflux-like Symptoms/Atypical Chest Pain\n   11
#> 105            Ives,Rashiah  Reflux-like Symptoms/Atypical Chest Pain\n    7
#> 106        Kekich,Annabelle  Reflux-like Symptoms/Atypical Chest Pain\n   12
#> 107   Kola Kehinde, Karisma  Reflux-like Symptoms/Atypical Chest Pain\n    9
#> 108         Martinez,Maegen  Reflux-like Symptoms/Atypical Chest Pain\n    7
#> 109           Moreno,Lauren  Reflux-like Symptoms/Atypical Chest Pain\n   13
#> 110         Sullivan,Shelby  Reflux-like Symptoms/Atypical Chest Pain\n   15
#> 111  al Arif, Ummu Kulthoom                        Small Bowel Biopsy\n    8
#> 112          Anderson,Alana                        Small Bowel Biopsy\n   13
#> 113 Avitia Ramirez, Alondra                        Small Bowel Biopsy\n   16
#> 114          Greimann,Phoua                        Small Bowel Biopsy\n    9
#> 115            Ives,Rashiah                        Small Bowel Biopsy\n   10
#> 116        Kekich,Annabelle                        Small Bowel Biopsy\n   12
#> 117   Kola Kehinde, Karisma                        Small Bowel Biopsy\n    7
#> 118         Martinez,Maegen                        Small Bowel Biopsy\n   14
#> 119           Moreno,Lauren                        Small Bowel Biopsy\n   12
#> 120         Sullivan,Shelby                        Small Bowel Biopsy\n    6
#> 121  al Arif, Ummu Kulthoom                    Surveillance-Barrett's\n   19
#> 122          Anderson,Alana                    Surveillance-Barrett's\n   13
#> 123 Avitia Ramirez, Alondra                    Surveillance-Barrett's\n   19
#> 124          Greimann,Phoua                    Surveillance-Barrett's\n   13
#> 125            Ives,Rashiah                    Surveillance-Barrett's\n   13
#> 126        Kekich,Annabelle                    Surveillance-Barrett's\n   11
#> 127   Kola Kehinde, Karisma                    Surveillance-Barrett's\n   10
#> 128         Martinez,Maegen                    Surveillance-Barrett's\n   10
#> 129           Moreno,Lauren                    Surveillance-Barrett's\n   17
#> 130         Sullivan,Shelby                    Surveillance-Barrett's\n   14
#> 131  al Arif, Ummu Kulthoom                   Therapeutic- Dilatation\n   10
#> 132          Anderson,Alana                   Therapeutic- Dilatation\n   11
#> 133 Avitia Ramirez, Alondra                   Therapeutic- Dilatation\n    9
#> 134          Greimann,Phoua                   Therapeutic- Dilatation\n   13
#> 135            Ives,Rashiah                   Therapeutic- Dilatation\n   18
#> 136        Kekich,Annabelle                   Therapeutic- Dilatation\n    9
#> 137   Kola Kehinde, Karisma                   Therapeutic- Dilatation\n   10
#> 138         Martinez,Maegen                   Therapeutic- Dilatation\n   16
#> 139           Moreno,Lauren                   Therapeutic- Dilatation\n   17
#> 140         Sullivan,Shelby                   Therapeutic- Dilatation\n    9
#> 141  al Arif, Ummu Kulthoom                          Therapeutic- RFA\n   12
#> 142          Anderson,Alana                          Therapeutic- RFA\n    9
#> 143 Avitia Ramirez, Alondra                          Therapeutic- RFA\n   11
#> 144          Greimann,Phoua                          Therapeutic- RFA\n    8
#> 145            Ives,Rashiah                          Therapeutic- RFA\n    8
#> 146        Kekich,Annabelle                          Therapeutic- RFA\n   16
#> 147   Kola Kehinde, Karisma                          Therapeutic- RFA\n   16
#> 148         Martinez,Maegen                          Therapeutic- RFA\n    8
#> 149           Moreno,Lauren                          Therapeutic- RFA\n   10
#> 150         Sullivan,Shelby                          Therapeutic- RFA\n   11
#> 151  al Arif, Ummu Kulthoom                               Weight Loss\n   19
#> 152          Anderson,Alana                               Weight Loss\n   11
#> 153 Avitia Ramirez, Alondra                               Weight Loss\n    9
#> 154          Greimann,Phoua                               Weight Loss\n   13
#> 155            Ives,Rashiah                               Weight Loss\n   11
#> 156        Kekich,Annabelle                               Weight Loss\n   15
#> 157   Kola Kehinde, Karisma                               Weight Loss\n   16
#> 158         Martinez,Maegen                               Weight Loss\n   14
#> 159           Moreno,Lauren                               Weight Loss\n   14
#> 160         Sullivan,Shelby                               Weight Loss\n    5

 rm(Myendo)