Determine polyp location

This should be used after the TermStandardizer as it relies on the presence of a SampleLocation column which the TermStandardLocation produces. It should be used after the PolypTidyUpLocator

PolypLocator(x, y)

Arguments

x
The dataframe

y
The column containing the SampleLocation from the TermStandardLocation

Examples

Histoltree <-list("Hospital Number:","Patient Name:",
"General Practitioner:","Date received:","Clinical Details",
 "Nature of specimen","Histology","Diagnosis",""
)

for (i in 1:(length(Histoltree) - 1)) {
 Mypath <-Extractor(
     Mypath, "PathReportWhole",as.character(Histoltree[i]),
     as.character(Histoltree[i + 1]),
     gsub(" ", "", as.character(Histoltree[i]))
   )
}

#> Error in Extractor(Mypath, "PathReportWhole", as.character(Histoltree[i]),     as.character(Histoltree[i + 1]), gsub(" ", "", as.character(Histoltree[i]))): could not find function "Extractor"

names(Mypath)[names(Mypath) == 'Datereceived'] <- 'Dateofprocedure'
Mypath$Dateofprocedure <- as.Date(Mypath$Dateofprocedure)
# The polyp locator then determines where the biopsies were taken from
# by assessing the SampleLocation column which comes from the
# TermStandardLocation function.
f<-TermStandardLocation(Mypath,'Histology')
f<-PolypLocator(f,'SampleLocation')
rm(Mypath)