Extract the last test done by a patient in days and how long ago

This determines the last test done by that patient and the time between now and that last test in days

SurveilLastToNow(dataframe, HospNum_Id, Endo_ResultPerformed)

Arguments

dataframe
dataframe

HospNum_Id
Patient ID

Endo_ResultPerformed
Date of the Endoscopy

Examples

SurveilLastToNow(Myendo,'HospitalNumber',
'Dateofprocedure')

#> # A tibble: 2,000 x 14
#> # Groups:   HospitalNumber [100]
#>    OGDReportWhole   HospitalNumber PatientName GeneralPractiti… Dateofprocedure
#>    <chr>            <chr>          <chr>       <chr>            <date>         
#>  1 "Hospital: Rand… " A1648588\n"  "  Pirkey,… " Dr\n el-Mogha… 2001-01-25     
#>  2 "Hospital: Rand… " A1648588\n"  "  Goodwat… " Dr\n Yohannes… 2002-06-30     
#>  3 "Hospital: Rand… " A1648588\n"  "  Ketchum… " Dr\n Kemp, Ba… 2002-12-17     
#>  4 "Hospital: Rand… " A1648588\n"  "  Hamm, J… " Dr\n Simmons,… 2004-01-21     
#>  5 "Hospital: Rand… " A1648588\n"  "  al-Zafa… " Dr\n Gavin, L… 2004-06-30     
#>  6 "Hospital: Rand… " A1648588\n"  "  el-Muha… " Dr\n Zewdu, B… 2005-11-22     
#>  7 "Hospital: Rand… " A1648588\n"  "  Kanchan… " Dr\n Meeker, … 2009-02-02     
#>  8 "Hospital: Rand… " A1648588\n"  "  Eshbaug… " Dr\n Pacht, M… 2009-05-16     
#>  9 "Hospital: Rand… " A1648588\n"  "  Pak, El… " Dr\n Perry, N… 2009-07-19     
#> 10 "Hospital: Rand… " A1648588\n"  "  Jamerso… " Dr\n al-Korom… 2009-11-15     
#> # ... with 1,990 more rows, and 9 more variables: Endoscopist <chr>,
#> #   Secondendoscopist <chr>, Medications <chr>, Instrument <chr>,
#> #   ExtentofExam <chr>, Indications <chr>, ProcedurePerformed <chr>,
#> #   Findings <chr>, diffDate <time>