CRIM status Barrett's

This collects the patients in whom it is assumed that complete clearance of intestinal metasplasia has occurred (CRIM) after ablation therapy. This is done by collecting those endoscopies where the last EVENT was equal to 'nothing' when the patient had undergone radiofrequency ablation at some point

Barretts_CRIM(x, HospNum, EVENT)

Arguments

x
The dataframe

HospNum
The Hospital Number column

EVENT
The column called EVENT that determines what procedure the patient had at that endoscopy

Examples

# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
v<-HistolChopperAccessionNumber(Mypath,'Histology',
'SP-\\d{2}-\\d{7}')
v<-HistolChopperDx(v,'Diagnosis')

#> Warning: provided 11 variables to replace 1 variables

v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')

#> Error: `PatientName.x` = PatientName must be a symbol or a string, not formula

# The function relies on the other Barrett's functions being run as well:
b1<-BarrettsDataAccord_Prague(v,'Findings')

#> Error in `[.data.frame`(x, , y): undefined columns selected

b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')
colnames(b3)[colnames(b3) == 'pHospitalNum'] <- 'HospitalNumber'
# The function groups the procedures by patient and then looks at those which
# have 'nothing' in the event column (which means biopsies only) which was
# preceded by radiofrequency ablation (RFA) so that these patients are
# labelled as having clearance of intestinal metaplasia
Barretts_CRIM(b3,'HospitalNumber',"EVENT")

#> # A tibble: 752 x 35
#> # Groups:   HospitalNumber [100]
#>                                                                     PathReportWhole
#>                                                                               <chr>
#>  1 "SP-22-4817508\nHospital: Random NHS Foundation Trust\nHospital Number: A1648588
#>  2 "SP-24-9461967\nHospital: Random NHS Foundation Trust\nHospital Number: A1648588
#>  3 "SP-77-2775965\nHospital: Random NHS Foundation Trust\nHospital Number: A1648588
#>  4 "SP-72-6139190\nHospital: Random NHS Foundation Trust\nHospital Number: A1648588
#>  5 "SP-83-6838873\nHospital: Random NHS Foundation Trust\nHospital Number: A1648588
#>  6 "SP-29-1590254\nHospital: Random NHS Foundation Trust\nHospital Number: A1648588
#>  7 "SP-79-1431892\nHospital: Random NHS Foundation Trust\nHospital Number: A3962018
#>  8 "SP-84-9745332\nHospital: Random NHS Foundation Trust\nHospital Number: A3962018
#>  9 "SP-96-7908356\nHospital: Random NHS Foundation Trust\nHospital Number: A3962018
#> 10 "SP-74-3934567\nHospital: Random NHS Foundation Trust\nHospital Number: A3962018
#> # ... with 742 more rows, and 34 more variables: HospitalNumber <chr>,
#> #   PatientName.x <chr>, DOB <chr>, GeneralPractitioner.x <chr>, Date.x <date>,
#> #   ClinicalDetails <chr>, Macroscopicdescription <chr>, Histology <chr>,
#> #   Diagnosis <chr>, Dx_Simplified <chr>, Cancer <chr>, Dysplasia <chr>,
#> #   GIST <chr>, NumbOfBx <dbl>, BxSize <dbl>, OGDReportWhole <chr>,
#> #   eHospitalNum <chr>, PatientName.y <chr>, GeneralPractitioner.y <chr>,
#> #   Date.y <date>, Endoscopist <chr>, Secondendoscopist <chr>,
#> #   Medications <chr>, Instrument <chr>, ExtentofExam <chr>, Indications <chr>,
#> #   ProcedurePerformed <chr>, Findings <chr>, Days <time>, CStage <dbl>,
#> #   MStage <dbl>, IMorNoIM <chr>, EVENT <chr>, ind <lgl>

rm(v)