Plot a metric by endoscopist

This takes any of the numerical metrics in the dataset and plots it by endoscopist. It of course relies on a Endoscopist column being present

MetricByEndoscopist(dataframe, Column, EndoscopistColumn)

Arguments

dataframe
The dataframe

Column
The column (numeric data) of interest

EndoscopistColumn
The endoscopist column

Examples

#The function plots any numeric metric by endoscopist
# and also gives a table with it. In this example we plot medication by
# endoscopist
Myendo<-EndoscMeds(Myendo,'Medications')
kk<-MetricByEndoscopist(Myendo,'Endoscopist','Fent')
rm(Myendo)