Plot a metric by endoscopist

This takes any numerical metric in the dataset and plots it by endoscopist. It of course relies on a Endoscopist column being present

MetricByEndoscopist(x, y, z)

Arguments

x
The dataframe

y
The column (numeric data) of interest

z
The endoscopist column

Examples

#The function plots any numeric metric by endoscopist
# and also gives a table with it. In this example we plot medication by
# endoscopist
Myendo<-EndoscChopperMeds(Myendo,'Medications')
Fent<-MetricByEndoscopist(Myendo,'Endoscopist','Fent')
rm(Myendo)