Determine overall number of procedures performed

Determines the number of endoscopies done by an endoscopist by type of endosopy and indication for a given timeframe As per BSG recommendations for Upper GI minimum number of gastroscopies in a year (although here the time frame is user defined)

NumberPerformed(x, y, z)

Arguments

x
The dataframe

y
The column containing the Endoscopists names

z
The column containing the Indication for the examination

Examples

Myendo<-Myendo[grepl('Gastroscopy',Myendo$ProcedurePerformed),]
 NumberPerformed(Myendo,'Endoscopist','Indications')

#>                                Var1                                        Var2
#> 1     Dr\n al-Arif, Ummu Kulthoom\n         chronic abdo pain and constipaton\n
#> 2            Dr\n Anderson, Alana\n         chronic abdo pain and constipaton\n
#> 3    Dr\n Avitia-Ramirez, Alondra\n         chronic abdo pain and constipaton\n
#> 4            Dr\n Greimann, Phoua\n         chronic abdo pain and constipaton\n
#> 5              Dr\n Ives, Rashiah\n         chronic abdo pain and constipaton\n
#> 6          Dr\n Kekich, Annabelle\n         chronic abdo pain and constipaton\n
#> 7      Dr\n Kola-Kehinde, Karisma\n         chronic abdo pain and constipaton\n
#> 8           Dr\n Martinez, Maegen\n         chronic abdo pain and constipaton\n
#> 9             Dr\n Moreno, Lauren\n         chronic abdo pain and constipaton\n
#> 10          Dr\n Sullivan, Shelby\n         chronic abdo pain and constipaton\n
#> 11    Dr\n al-Arif, Ummu Kulthoom\n                                 Dyspepsia\n
#> 12           Dr\n Anderson, Alana\n                                 Dyspepsia\n
#> 13   Dr\n Avitia-Ramirez, Alondra\n                                 Dyspepsia\n
#> 14           Dr\n Greimann, Phoua\n                                 Dyspepsia\n
#> 15             Dr\n Ives, Rashiah\n                                 Dyspepsia\n
#> 16         Dr\n Kekich, Annabelle\n                                 Dyspepsia\n
#> 17     Dr\n Kola-Kehinde, Karisma\n                                 Dyspepsia\n
#> 18          Dr\n Martinez, Maegen\n                                 Dyspepsia\n
#> 19            Dr\n Moreno, Lauren\n                                 Dyspepsia\n
#> 20          Dr\n Sullivan, Shelby\n                                 Dyspepsia\n
#> 21    Dr\n al-Arif, Ummu Kulthoom\n   Dysphagia/intermittent for a few months\n
#> 22           Dr\n Anderson, Alana\n   Dysphagia/intermittent for a few months\n
#> 23   Dr\n Avitia-Ramirez, Alondra\n   Dysphagia/intermittent for a few months\n
#> 24           Dr\n Greimann, Phoua\n   Dysphagia/intermittent for a few months\n
#> 25             Dr\n Ives, Rashiah\n   Dysphagia/intermittent for a few months\n
#> 26         Dr\n Kekich, Annabelle\n   Dysphagia/intermittent for a few months\n
#> 27     Dr\n Kola-Kehinde, Karisma\n   Dysphagia/intermittent for a few months\n
#> 28          Dr\n Martinez, Maegen\n   Dysphagia/intermittent for a few months\n
#> 29            Dr\n Moreno, Lauren\n   Dysphagia/intermittent for a few months\n
#> 30          Dr\n Sullivan, Shelby\n   Dysphagia/intermittent for a few months\n
#> 31    Dr\n al-Arif, Ummu Kulthoom\n                     Dysphagia/Odynophagia\n
#> 32           Dr\n Anderson, Alana\n                     Dysphagia/Odynophagia\n
#> 33   Dr\n Avitia-Ramirez, Alondra\n                     Dysphagia/Odynophagia\n
#> 34           Dr\n Greimann, Phoua\n                     Dysphagia/Odynophagia\n
#> 35             Dr\n Ives, Rashiah\n                     Dysphagia/Odynophagia\n
#> 36         Dr\n Kekich, Annabelle\n                     Dysphagia/Odynophagia\n
#> 37     Dr\n Kola-Kehinde, Karisma\n                     Dysphagia/Odynophagia\n
#> 38          Dr\n Martinez, Maegen\n                     Dysphagia/Odynophagia\n
#> 39            Dr\n Moreno, Lauren\n                     Dysphagia/Odynophagia\n
#> 40          Dr\n Sullivan, Shelby\n                     Dysphagia/Odynophagia\n
#> 41    Dr\n al-Arif, Ummu Kulthoom\n                   Follow-up ULCER HEALING\n
#> 42           Dr\n Anderson, Alana\n                   Follow-up ULCER HEALING\n
#> 43   Dr\n Avitia-Ramirez, Alondra\n                   Follow-up ULCER HEALING\n
#> 44           Dr\n Greimann, Phoua\n                   Follow-up ULCER HEALING\n
#> 45             Dr\n Ives, Rashiah\n                   Follow-up ULCER HEALING\n
#> 46         Dr\n Kekich, Annabelle\n                   Follow-up ULCER HEALING\n
#> 47     Dr\n Kola-Kehinde, Karisma\n                   Follow-up ULCER HEALING\n
#> 48          Dr\n Martinez, Maegen\n                   Follow-up ULCER HEALING\n
#> 49            Dr\n Moreno, Lauren\n                   Follow-up ULCER HEALING\n
#> 50          Dr\n Sullivan, Shelby\n                   Follow-up ULCER HEALING\n
#> 51    Dr\n al-Arif, Ummu Kulthoom\n          Haematemesis or Melaena/Blood PR\n
#> 52           Dr\n Anderson, Alana\n          Haematemesis or Melaena/Blood PR\n
#> 53   Dr\n Avitia-Ramirez, Alondra\n          Haematemesis or Melaena/Blood PR\n
#> 54           Dr\n Greimann, Phoua\n          Haematemesis or Melaena/Blood PR\n
#> 55             Dr\n Ives, Rashiah\n          Haematemesis or Melaena/Blood PR\n
#> 56         Dr\n Kekich, Annabelle\n          Haematemesis or Melaena/Blood PR\n
#> 57     Dr\n Kola-Kehinde, Karisma\n          Haematemesis or Melaena/Blood PR\n
#> 58          Dr\n Martinez, Maegen\n          Haematemesis or Melaena/Blood PR\n
#> 59            Dr\n Moreno, Lauren\n          Haematemesis or Melaena/Blood PR\n
#> 60          Dr\n Sullivan, Shelby\n          Haematemesis or Melaena/Blood PR\n
#> 61    Dr\n al-Arif, Ummu Kulthoom\n                    Nausea and/or Vomiting\n
#> 62           Dr\n Anderson, Alana\n                    Nausea and/or Vomiting\n
#> 63   Dr\n Avitia-Ramirez, Alondra\n                    Nausea and/or Vomiting\n
#> 64           Dr\n Greimann, Phoua\n                    Nausea and/or Vomiting\n
#> 65             Dr\n Ives, Rashiah\n                    Nausea and/or Vomiting\n
#> 66         Dr\n Kekich, Annabelle\n                    Nausea and/or Vomiting\n
#> 67     Dr\n Kola-Kehinde, Karisma\n                    Nausea and/or Vomiting\n
#> 68          Dr\n Martinez, Maegen\n                    Nausea and/or Vomiting\n
#> 69            Dr\n Moreno, Lauren\n                    Nausea and/or Vomiting\n
#> 70          Dr\n Sullivan, Shelby\n                    Nausea and/or Vomiting\n
#> 71    Dr\n al-Arif, Ummu Kulthoom\n                     Oesophagus- Dysplasia\n
#> 72           Dr\n Anderson, Alana\n                     Oesophagus- Dysplasia\n
#> 73   Dr\n Avitia-Ramirez, Alondra\n                     Oesophagus- Dysplasia\n
#> 74           Dr\n Greimann, Phoua\n                     Oesophagus- Dysplasia\n
#> 75             Dr\n Ives, Rashiah\n                     Oesophagus- Dysplasia\n
#> 76         Dr\n Kekich, Annabelle\n                     Oesophagus- Dysplasia\n
#> 77     Dr\n Kola-Kehinde, Karisma\n                     Oesophagus- Dysplasia\n
#> 78          Dr\n Martinez, Maegen\n                     Oesophagus- Dysplasia\n
#> 79            Dr\n Moreno, Lauren\n                     Oesophagus- Dysplasia\n
#> 80          Dr\n Sullivan, Shelby\n                     Oesophagus- Dysplasia\n
#> 81    Dr\n al-Arif, Ummu Kulthoom\n                                    Other-\n
#> 82           Dr\n Anderson, Alana\n                                    Other-\n
#> 83   Dr\n Avitia-Ramirez, Alondra\n                                    Other-\n
#> 84           Dr\n Greimann, Phoua\n                                    Other-\n
#> 85             Dr\n Ives, Rashiah\n                                    Other-\n
#> 86         Dr\n Kekich, Annabelle\n                                    Other-\n
#> 87     Dr\n Kola-Kehinde, Karisma\n                                    Other-\n
#> 88          Dr\n Martinez, Maegen\n                                    Other-\n
#> 89            Dr\n Moreno, Lauren\n                                    Other-\n
#> 90          Dr\n Sullivan, Shelby\n                                    Other-\n
#> 91    Dr\n al-Arif, Ummu Kulthoom\n               Previous OGD ? 8 months ago\n
#> 92           Dr\n Anderson, Alana\n               Previous OGD ? 8 months ago\n
#> 93   Dr\n Avitia-Ramirez, Alondra\n               Previous OGD ? 8 months ago\n
#> 94           Dr\n Greimann, Phoua\n               Previous OGD ? 8 months ago\n
#> 95             Dr\n Ives, Rashiah\n               Previous OGD ? 8 months ago\n
#> 96         Dr\n Kekich, Annabelle\n               Previous OGD ? 8 months ago\n
#> 97     Dr\n Kola-Kehinde, Karisma\n               Previous OGD ? 8 months ago\n
#> 98          Dr\n Martinez, Maegen\n               Previous OGD ? 8 months ago\n
#> 99            Dr\n Moreno, Lauren\n               Previous OGD ? 8 months ago\n
#> 100         Dr\n Sullivan, Shelby\n               Previous OGD ? 8 months ago\n
#> 101   Dr\n al-Arif, Ummu Kulthoom\n  Reflux-like Symptoms/Atypical Chest Pain\n
#> 102          Dr\n Anderson, Alana\n  Reflux-like Symptoms/Atypical Chest Pain\n
#> 103  Dr\n Avitia-Ramirez, Alondra\n  Reflux-like Symptoms/Atypical Chest Pain\n
#> 104          Dr\n Greimann, Phoua\n  Reflux-like Symptoms/Atypical Chest Pain\n
#> 105            Dr\n Ives, Rashiah\n  Reflux-like Symptoms/Atypical Chest Pain\n
#> 106        Dr\n Kekich, Annabelle\n  Reflux-like Symptoms/Atypical Chest Pain\n
#> 107    Dr\n Kola-Kehinde, Karisma\n  Reflux-like Symptoms/Atypical Chest Pain\n
#> 108         Dr\n Martinez, Maegen\n  Reflux-like Symptoms/Atypical Chest Pain\n
#> 109           Dr\n Moreno, Lauren\n  Reflux-like Symptoms/Atypical Chest Pain\n
#> 110         Dr\n Sullivan, Shelby\n  Reflux-like Symptoms/Atypical Chest Pain\n
#> 111   Dr\n al-Arif, Ummu Kulthoom\n                        Small Bowel Biopsy\n
#> 112          Dr\n Anderson, Alana\n                        Small Bowel Biopsy\n
#> 113  Dr\n Avitia-Ramirez, Alondra\n                        Small Bowel Biopsy\n
#> 114          Dr\n Greimann, Phoua\n                        Small Bowel Biopsy\n
#> 115            Dr\n Ives, Rashiah\n                        Small Bowel Biopsy\n
#> 116        Dr\n Kekich, Annabelle\n                        Small Bowel Biopsy\n
#> 117    Dr\n Kola-Kehinde, Karisma\n                        Small Bowel Biopsy\n
#> 118         Dr\n Martinez, Maegen\n                        Small Bowel Biopsy\n
#> 119           Dr\n Moreno, Lauren\n                        Small Bowel Biopsy\n
#> 120         Dr\n Sullivan, Shelby\n                        Small Bowel Biopsy\n
#> 121   Dr\n al-Arif, Ummu Kulthoom\n                    Surveillance-Barrett's\n
#> 122          Dr\n Anderson, Alana\n                    Surveillance-Barrett's\n
#> 123  Dr\n Avitia-Ramirez, Alondra\n                    Surveillance-Barrett's\n
#> 124          Dr\n Greimann, Phoua\n                    Surveillance-Barrett's\n
#> 125            Dr\n Ives, Rashiah\n                    Surveillance-Barrett's\n
#> 126        Dr\n Kekich, Annabelle\n                    Surveillance-Barrett's\n
#> 127    Dr\n Kola-Kehinde, Karisma\n                    Surveillance-Barrett's\n
#> 128         Dr\n Martinez, Maegen\n                    Surveillance-Barrett's\n
#> 129           Dr\n Moreno, Lauren\n                    Surveillance-Barrett's\n
#> 130         Dr\n Sullivan, Shelby\n                    Surveillance-Barrett's\n
#> 131   Dr\n al-Arif, Ummu Kulthoom\n                   Therapeutic- Dilatation\n
#> 132          Dr\n Anderson, Alana\n                   Therapeutic- Dilatation\n
#> 133  Dr\n Avitia-Ramirez, Alondra\n                   Therapeutic- Dilatation\n
#> 134          Dr\n Greimann, Phoua\n                   Therapeutic- Dilatation\n
#> 135            Dr\n Ives, Rashiah\n                   Therapeutic- Dilatation\n
#> 136        Dr\n Kekich, Annabelle\n                   Therapeutic- Dilatation\n
#> 137    Dr\n Kola-Kehinde, Karisma\n                   Therapeutic- Dilatation\n
#> 138         Dr\n Martinez, Maegen\n                   Therapeutic- Dilatation\n
#> 139           Dr\n Moreno, Lauren\n                   Therapeutic- Dilatation\n
#> 140         Dr\n Sullivan, Shelby\n                   Therapeutic- Dilatation\n
#> 141   Dr\n al-Arif, Ummu Kulthoom\n                          Therapeutic- RFA\n
#> 142          Dr\n Anderson, Alana\n                          Therapeutic- RFA\n
#> 143  Dr\n Avitia-Ramirez, Alondra\n                          Therapeutic- RFA\n
#> 144          Dr\n Greimann, Phoua\n                          Therapeutic- RFA\n
#> 145            Dr\n Ives, Rashiah\n                          Therapeutic- RFA\n
#> 146        Dr\n Kekich, Annabelle\n                          Therapeutic- RFA\n
#> 147    Dr\n Kola-Kehinde, Karisma\n                          Therapeutic- RFA\n
#> 148         Dr\n Martinez, Maegen\n                          Therapeutic- RFA\n
#> 149           Dr\n Moreno, Lauren\n                          Therapeutic- RFA\n
#> 150         Dr\n Sullivan, Shelby\n                          Therapeutic- RFA\n
#> 151   Dr\n al-Arif, Ummu Kulthoom\n                               Weight Loss\n
#> 152          Dr\n Anderson, Alana\n                               Weight Loss\n
#> 153  Dr\n Avitia-Ramirez, Alondra\n                               Weight Loss\n
#> 154          Dr\n Greimann, Phoua\n                               Weight Loss\n
#> 155            Dr\n Ives, Rashiah\n                               Weight Loss\n
#> 156        Dr\n Kekich, Annabelle\n                               Weight Loss\n
#> 157    Dr\n Kola-Kehinde, Karisma\n                               Weight Loss\n
#> 158         Dr\n Martinez, Maegen\n                               Weight Loss\n
#> 159           Dr\n Moreno, Lauren\n                               Weight Loss\n
#> 160         Dr\n Sullivan, Shelby\n                               Weight Loss\n
#>     Freq
#> 1     17
#> 2     10
#> 3     10
#> 4     12
#> 5     16
#> 6     11
#> 7     10
#> 8     13
#> 9     15
#> 10    13
#> 11    18
#> 12    13
#> 13     8
#> 14    11
#> 15    15
#> 16    12
#> 17    16
#> 18    12
#> 19    18
#> 20    17
#> 21    10
#> 22    20
#> 23     5
#> 24    14
#> 25    13
#> 26     7
#> 27    13
#> 28    12
#> 29    11
#> 30    13
#> 31    13
#> 32    13
#> 33     8
#> 34    11
#> 35    15
#> 36    12
#> 37     7
#> 38    12
#> 39    19
#> 40     9
#> 41     4
#> 42    11
#> 43    16
#> 44    12
#> 45    12
#> 46    15
#> 47    14
#> 48     9
#> 49    10
#> 50    15
#> 51    17
#> 52    13
#> 53     9
#> 54     7
#> 55     9
#> 56     4
#> 57    13
#> 58     6
#> 59    12
#> 60    11
#> 61     9
#> 62     7
#> 63    13
#> 64     7
#> 65    13
#> 66     6
#> 67     7
#> 68    11
#> 69    10
#> 70     7
#> 71     7
#> 72    18
#> 73     8
#> 74     8
#> 75    17
#> 76    13
#> 77    19
#> 78    12
#> 79    11
#> 80    16
#> 81    32
#> 82    24
#> 83    23
#> 84    21
#> 85    21
#> 86    19
#> 87    20
#> 88    25
#> 89    18
#> 90    18
#> 91    18
#> 92     9
#> 93    18
#> 94    11
#> 95    12
#> 96     7
#> 97     9
#> 98    12
#> 99    15
#> 100   14
#> 101    9
#> 102   14
#> 103   11
#> 104   11
#> 105    7
#> 106   12
#> 107    9
#> 108    7
#> 109   13
#> 110   15
#> 111    8
#> 112   13
#> 113   16
#> 114    9
#> 115   10
#> 116   12
#> 117    7
#> 118   14
#> 119   12
#> 120    6
#> 121   19
#> 122   13
#> 123   19
#> 124   13
#> 125   13
#> 126   11
#> 127   10
#> 128   10
#> 129   17
#> 130   14
#> 131   10
#> 132   11
#> 133    9
#> 134   13
#> 135   18
#> 136    9
#> 137   10
#> 138   16
#> 139   17
#> 140    9
#> 141   12
#> 142    9
#> 143   11
#> 144    8
#> 145    8
#> 146   16
#> 147   16
#> 148    8
#> 149   10
#> 150   11
#> 151   19
#> 152   11
#> 153    9
#> 154   13
#> 155   11
#> 156   15
#> 157   16
#> 158   14
#> 159   14
#> 160    5